Rename BridgeStorageAccessor to StorageAccessor. Make StorageAccessor, FeePerKbStorageProvider and FeePerKbSupport interfaces and create their implementations
